Rename worker-adapter dir to adapters and bless v2 API baseline
Final chunk of the v2 extraction/loading refactor.

- Move src/multithreading/worker-adapter/ -> src/multithreading/adapters/
  and drop the worker- prefix on test files, naming each after its
  subject (base-adapter.emit, extraction-adapter, loading-adapter, ...).
  The adapters are the dev-facing API, not thread plumbing, so they no
  longer live under a "worker" folder name. All import paths updated.
- Regenerate the backwards-compatibility baseline (latest.json +
  airsync-sdk.api.md) to bless the v2 public API: processExtractionTask /
  processLoadingTask / ExtractionAdapter / LoadingAdapter replace
  processTask / WorkerAdapter, and AdapterState is gone.

All suites green: default, timeout-handling, slow, backwards-compatibility.
