+++ src/multithreading/adapters/base-adapter.emit.test.ts
