+++ src/multithreading/adapters/base-adapter.ts
