+++ src/multithreading/adapters/extraction-adapter.artifacts.test.ts
