+++ src/multithreading/adapters/extraction-adapter.test.ts
