+++ src/multithreading/adapters/extraction-adapter.ts
