+++ src/multithreading/adapters/loading-adapter.helpers.ts
