+++ src/multithreading/adapters/loading-adapter.ts
