+++ src/multithreading/adapters/serialization.test.ts
